TASK: Rename NoChangesException to WorkspaceCommandSkipped
Instead of keeping it specific to publishing and naming it `NoPublishableChanges` we use it as generic no-op exception for workspace commands. Especially because `becauseWorkspaceToRebaseIsNotOutdated` already deviates from the 'PublishableChanges' kind.
