+++ Neos.ContentRepository.BehavioralTests/Tests/Behavior/Features/W6-WorkspaceRebasing/01-BasicFeatures.feature
@@ -46,7 +46,7 @@ Feature: Rebasing with no conflict
       | Key                         | Value                 |
       | workspaceName               | "user-test"           |
       | rebasedContentStreamId      | "user-cs-rebased"     |
-    Then the last command should have thrown an exception of type "NoChangesException" with code 1730463693 and message:
+    Then the last command should have thrown an exception of type "WorkspaceCommandSkipped" with code 1730463693 and message:
     """
     Skipped rebase workspace "user-test" because it is not outdated.
     """
